+++ packages/expo-image/android/src/main/java/expo/modules/image/blurhash/BlurhashModelLoaderFactory.kt
@@ -4,10 +4,10 @@ import android.graphics.Bitmap
 import com.bumptech.glide.load.model.ModelLoader
 import com.bumptech.glide.load.model.ModelLoaderFactory
 import com.bumptech.glide.load.model.MultiModelLoaderFactory
-import expo.modules.image.GlideBlurhashModel
+import expo.modules.image.BlurhashModelProvider
 
-class BlurhashModelLoaderFactory : ModelLoaderFactory<GlideBlurhashModel, Bitmap> {
-  override fun build(multiFactory: MultiModelLoaderFactory): ModelLoader<GlideBlurhashModel, Bitmap> =
+class BlurhashModelLoaderFactory : ModelLoaderFactory<BlurhashModelProvider, Bitmap> {
+  override fun build(multiFactory: MultiModelLoaderFactory): ModelLoader<BlurhashModelProvider, Bitmap> =
     BlurhashModelLoader()
 
   override fun teardown() = Unit