+++ packages/expo-image/android/src/main/java/expo/modules/image/decodedsource/DecodedModule.kt
@@ -1,17 +1,16 @@
-package expo.modules.image.thumbhash
+package expo.modules.image.decodedsource
 
 import android.content.Context
-import android.graphics.Bitmap
+import android.graphics.drawable.Drawable
 import com.bumptech.glide.Glide
 import com.bumptech.glide.Registry
 import com.bumptech.glide.annotation.GlideModule
 import com.bumptech.glide.module.LibraryGlideModule
-import expo.modules.image.ThumbhashModelProvider
 
 @GlideModule
-class ThumbhashModule : LibraryGlideModule() {
+class DecodedModule : LibraryGlideModule() {
   override fun registerComponents(context: Context, glide: Glide, registry: Registry) {
     super.registerComponents(context, glide, registry)
-    registry.prepend(ThumbhashModelProvider::class.java, Bitmap::class.java, ThumbhashModelLoaderFactory())
+    registry.prepend(DecodedModel::class.java, Drawable::class.java, DecodedModelLoaderFactory())
   }
 }