+++ packages/expo-image/android/src/main/java/expo/modules/image/thumbhash/ThumbhashModelLoaderFactory.kt
@@ -4,10 +4,10 @@ import android.graphics.Bitmap
 import com.bumptech.glide.load.model.ModelLoader
 import com.bumptech.glide.load.model.ModelLoaderFactory
 import com.bumptech.glide.load.model.MultiModelLoaderFactory
-import expo.modules.image.GlideThumbhashModel
+import expo.modules.image.ThumbhashModelProvider
 
-class ThumbhashModelLoaderFactory : ModelLoaderFactory<GlideThumbhashModel, Bitmap> {
-  override fun build(multiFactory: MultiModelLoaderFactory): ModelLoader<GlideThumbhashModel, Bitmap> =
+class ThumbhashModelLoaderFactory : ModelLoaderFactory<ThumbhashModelProvider, Bitmap> {
+  override fun build(multiFactory: MultiModelLoaderFactory): ModelLoader<ThumbhashModelProvider, Bitmap> =
     ThumbhashModelLoader()
 
   override fun teardown() = Unit